Fix build docker with sqlite3 musl error
This is now broken in latest `golang:alpine` image (from alpine3.18 to alpine3.19).
Alpine 3.19 was released on 2023-12-07. [^1] It is shipped with gcc 13.2 and musl-dev 2.14. [^2]
The LFS64 interfaces were marked as deprecated for musl 1.2.4. [^3]

As a temporary workaround, builds broken by this change can be fixed short-term by adding `-D_LARGEFILE64_SOURCE` to `CFLAGS`, but should be fixed to use the standard interfaces, refer to go-sqlite3 issue 1164. [^4]

---

[^1]: https://www.alpinelinux.org/posts/Alpine-3.19.0-released.html
[^2]: https://pkgs.alpinelinux.org/packages?name=musl-dev&branch=v3.19
[^3]: https://git.musl-libc.org/cgit/musl/commit/?id=25e6fee27f4a293728dd15b659170e7b9c7db9bc
[^4]: https://github.com/mattn/go-sqlite3/issues/1164