fix: missing version information for command
The version information is missing when running `go-xn version` or `go-xn --version`.
So update variable name in `LDFLAGS` to match the variable name in `cmd` package.

The bug was introduced in PR #368 , which updated the variable name in the `cmd` package from `Version` to `version`, but the `Makefile` was not updated accordingly.