Add support for funding manifest and discovery
Add static route for `/.well-known/funding-manifest-urls` and `/funding.json` file from multiple file systems.

`funding.json` is an open manifest (JSON schema) that acts as a signaling and discovery mechanism for projects seeking funding.
`funding-manifest-urls` is a text file that contains URL (one or more) pointing to the `funding.json` file, which follows the `.well-known` convention.

refer to:
- https://floss.fund/funding-manifest/
- https://datatracker.ietf.org/doc/html/rfc8615