Add `TelegramWriter` to send logs to telegram bot API
Implement a new `TelegramWriter` that complies with the `io.Writer` interface for the `slog` package.
It refers to the `TelegramHook` logrus hook in `github.com/Pengxn/go-xn/src/util/log/hook` package, but is designed for the `slog` package.
This writer will send logs to telegram using the telegram bot API, refer to [Telegram Bot API - sendMessage].

[Telegram Bot API - sendMessage]: https://core.telegram.org/bots/api#sendmessage