Refactor `TelegramHook` logrus hook to `TelegramWriter` (#439)
* Add `TelegramWriter` to send logs to telegram bot API

Implement a new `TelegramWriter` that complies with the `io.Writer` interface for the `slog` package.
It refers to the `TelegramHook` logrus hook in `github.com/Pengxn/go-xn/src/util/log/hook` package, but is designed for the `slog` package.
This writer will send logs to telegram using the telegram bot API, refer to [Telegram Bot API - sendMessage].

[Telegram Bot API - sendMessage]: https://core.telegram.org/bots/api#sendmessage

* Optimize error value by removing unnecessary wrapping of messages

- Fix variable shadowing by reusing the existing `err` variable in the `Write()` method.
- Remove unnecessary error wrapping in favor of direct error returns, as handled by the caller.

* Reduce variable assignments in formatting the message string

- Remove `serialized` variable by using `raw` parameter directly.
- Eliminate redundant `levelWithEmoji` variable by including the value directly in the format string.